+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Added MongoDB in docker compose and updated the backend to use the instance
 - ChatBot interface communicating with the OpenStack Ollama instance
 
 ---